makes acid and pacid respect new IMMUNE_TO_ACID flag. updates syndie_clown gas mask to use flag.

+++ code/modules/chemistry/Reagents-PoisonEtc.dm
@@ -112,11 +112,13 @@ datum
 				if (istype(O,/obj/item/clothing/head/chemhood || /obj/item/clothing/suit/chemsuit))
 					return 1
 				if (isitem(O) && prob(40))
-					var/obj/decal/cleanable/molten_item/I = make_cleanable(/obj/decal/cleanable/molten_item,O.loc)
-					I.desc = "Looks like this was \an [O] some time ago."
-					for(var/mob/M in AIviewers(5, O))
-						boutput(M, "<span class='alert'>\the [O] melts.</span>")
-					qdel(O)
+					var/obj/item/toMelt
+					if (!(toMelt.item_function_flags & IMMUNE_TO_ACID))
+						var/obj/decal/cleanable/molten_item/I = make_cleanable(/obj/decal/cleanable/molten_item,O.loc)
+						I.desc = "Looks like this was \an [O] some time ago."
+						for(var/mob/M in AIviewers(5, O))
+							boutput(M, "<span class='alert'>\the [O] melts.</span>")
+						qdel(O)
 
 			on_plant_life(var/obj/machinery/plantpot/P)
 				P.HYPdamageplant("acid",5)
@@ -791,23 +793,28 @@ datum
 							H.unlock_medal("Red Hood", 1)
 							return
 						else
-							if (H.wear_mask)
-								var/obj/item/clothing/mask/K = H.wear_mask
-								if ( !K.acid_proof )
-									boutput(M, "<span class='alert'>Your [H.wear_mask] melts away!</span>")
-									K.dropped(H)
-									H.u_equip(K)
-									qdel(K)
-								else
-									boutput(M, "<span class='alert'>Your [H.wear_mask] protects you from the acid!</span>")
-								melted = 1
 							if (H.head)
-								boutput(M, "<span class='alert'>Your [H.head] melts into uselessness!</span>")
 								var/obj/item/clothing/head/D = H.head
-								D.dropped(H)
-								H.u_equip(D)
-								qdel(D)
+								if (!(D.item_function_flags & IMMUNE_TO_ACID))
+									boutput(M, "<span class='alert'>Your [H.head] melts into uselessness!</span>")
+									D.dropped(H)
+									H.u_equip(D)
+									qdel(D)
+								else
+									boutput(M, "<span class='alert'>Your [H.head] protects you from the acid!</span>")
 								melted = 1
+							if (!(H?.head.c_flags & SPACEWEAR))
+								if (H.wear_mask)
+									var/obj/item/clothing/mask/K = H.wear_mask
+									if (!(K.item_function_flags & IMMUNE_TO_ACID))
+										boutput(M, "<span class='alert'>Your [H.wear_mask] melts away!</span>")
+										K.dropped(H)
+										H.u_equip(K)
+										qdel(K)
+									else
+										boutput(M, "<span class='alert'>Your [H.wear_mask] protects you from the acid!</span>")
+									melted = 1
+
 							if (melted)
 								return
 					else
@@ -819,7 +826,7 @@ datum
 						if (H.wear_mask && H.head)
 							if (H.wear_mask)
 								var/obj/item/clothing/mask/K = H.wear_mask
-								if ( !K.acid_proof )
+								if (!(K.item_function_flags & IMMUNE_TO_ACID))
 									boutput(M, "<span class='alert'>Your [H.wear_mask] melts away!</span>")
 									K.dropped(H)
 									H.u_equip(K)
@@ -828,11 +835,14 @@ datum
 									boutput(M, "<span class='alert'>Your [H.wear_mask] protects you from the acid!</span>")
 								melted = 1
 							if (H.head)
-								boutput(M, "<span class='alert'>Your [H.head] melts into uselessness!</span>")
 								var/obj/item/clothing/head/D = H.head
-								D.dropped(H)
-								H.u_equip(D)
-								qdel(D)
+								if (!(D.item_function_flags & IMMUNE_TO_ACID))
+									boutput(M, "<span class='alert'>Your [H.head] melts into uselessness!</span>")
+									D.dropped(H)
+									H.u_equip(D)
+									qdel(D)
+								else
+									boutput(M, "<span class='alert'>Your [H.head] protects you from the acid!</span>")
 								melted = 1
 							if (melted)
 								return
@@ -851,11 +861,13 @@ datum
 				if (istype(O,/obj/fluid))
 					return 1
 				if (isitem(O) && volume > O:w_class)
-					var/obj/decal/cleanable/molten_item/I = make_cleanable(/obj/decal/cleanable/molten_item,O.loc)
-					I.desc = "Looks like this was \an [O] some time ago."
-					for(var/mob/M in AIviewers(5, O))
-						boutput(M, "<span class='alert'>\the [O] melts.</span>")
-					qdel(O)
+					var/obj/item/toMelt = O
+					if (!(toMelt.item_function_flags & IMMUNE_TO_ACID))
+						var/obj/decal/cleanable/molten_item/I = make_cleanable(/obj/decal/cleanable/molten_item,O.loc)
+						I.desc = "Looks like this was \an [O] some time ago."
+						for(var/mob/M in AIviewers(5, O))
+							boutput(M, "<span class='alert'>\the [O] melts.</span>")
+						qdel(O)
 
 			on_plant_life(var/obj/machinery/plantpot/P)
 				P.HYPdamageplant("acid",10)
